tmp/requirements.txt
asgiref==3.6.0
attrs==22.1.0
beautifulsoup4==4.11.2
bootstrap4==0.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coverage==7.1.0
crispy-bootstrap4==2022.1
datadispatch==1.0.0
djando==0.2
Django==4.1.7
django-bootstrap4==22.3
django-bootstrap4-form==4.0.1
django-cleanup==7.0.0
django-crispy-forms==2.0
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
dominate==2.7.0
exceptiongroup==1.0.4
Faker==6.0.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
idna==3.3
importlib-metadata==5.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mixer==7.2.2
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
pyTelegramBotAPI==4.9.0
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tasks==2.8.0
text-unidecode==1.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.9
userapp==1.0.5
visitor==0.1.3
Werkzeug==2.2.2
zipp==3.14.0
